ReplayV2 fully functional :D
