+++ Replays/NewReplaySystem.cs
@@ -9,7 +9,6 @@ using Newtonsoft.Json;
 using TootTally.GameplayModifier;
 using TootTally.Graphics;
 using TootTally.Utils;
-using TootTally.Utils.APIServices;
 using TootTally.Utils.Helpers;
 using TrombLoader.CustomTracks;
 using UnityEngine;
@@ -18,177 +17,134 @@ namespace TootTally.Replays
 {
     public class NewReplaySystem
     {
-        public static List<string> incompatibleReplayPluginBuildDate = new List<string> { "20230106" };
-        public const string REPLAY_VERSION = "1.0.0";
+        public static List<string> incompatibleReplayVersions = new List<string> { "1.0.0" };
+        public const string REPLAY_VERSION = "2.0.0";
 
-        private int _scores_A, _scores_B, _scores_C, _scores_D, _scores_F, _totalScore, _finalTotalScore;
-        private int[] _finalNoteTally, _replayNoteTally; // [nasties, mehs, okays, nices, perfects]
         private int _frameIndex, _tootIndex;
-        private int _maxCombo;
-        private int _replaybuilddate;
 
-        private List<int[]> _frameData = new List<int[]>(), _noteData = new List<int[]>(), _tootData = new List<int[]>();
-        private int[] _lastFrameData;
-        private float _replaySpeed;
-        public float GetReplaySpeed { get => _replaySpeed; }
-        private DateTimeOffset _startTime, _endTime;
+        private ReplayData _replayData;
+        private dynamic[] _lastFrame, _currentFrame;
+        private dynamic[] _currentToot;
+        private dynamic[] _currentNote;
+
+        public float GetReplaySpeed { get => _replayData.gamespeedmultiplier; }
 
         private bool _wasTouchScreenUsed;
         private bool _wasTabletUsed;
         private bool _isTooting;
+        private int _maxCombo;
+        private bool _isLastNote;
         public bool GetIsTooting { get => _isTooting; }
 
-        private float _nextPositionTarget, _lastPosition;
-        private float _nextTimingTarget, _lastTiming;
-        private string _replayUsername, _replaySong;
-        public string GetUsername { get => _replayUsername; }
-        public string GetSongName { get => _replaySong; }
+        public string GetUsername { get => _replayData.username; }
+        public string GetSongName { get => _replayData.song; }
 
         public bool IsFullCombo { get => GlobalVariables.gameplay_notescores[0] == 0 && GlobalVariables.gameplay_notescores[1] == 0 && GlobalVariables.gameplay_notescores[2] == 0; }
         public bool IsTripleS { get => GlobalVariables.gameplay_notescores[0] == 0 && GlobalVariables.gameplay_notescores[1] == 0 && GlobalVariables.gameplay_notescores[2] == 0 && GlobalVariables.gameplay_notescores[3] == 0; }
 
+        public NewReplaySystem()
+        {
+            _replayData = new ReplayData(60);
+        }
+
         #region ReplayRecorder
 
-        public void SetupRecording(GameController __instance)
+        public void SetupRecording(int targetFramerate)
         {
-            _scores_A = _scores_B = _scores_C = _scores_D = 0;
-            _maxCombo = 0;
-            _lastFrameData = new int[4];
-
+            _replayData = new ReplayData(targetFramerate);
+            _replayData.gamemodifiers = GameModifierManager.GetModifiersString();
+            _replayData.scrollspeed = GlobalVariables.gamescrollspeed;
+            _replayData.gamespeedmultiplier = float.Parse(ReplaySystemManager.gameSpeedMultiplier.ToString("0.00"));
+            var track = TrackLookup.lookup(GlobalVariables.chosen_track_data.trackref);
+            _replayData.song = track.trackname_long;
+            if (track is CustomTrack)
+                _replayData.songhash = SongDataHelper.GetSongHash(track);
+            else
+                _replayData.songhash = SongDataHelper.CalcSHA256Hash(Encoding.UTF8.GetBytes(SongDataHelper.GenerateBaseTmb(track)));
             TootTallyLogger.LogInfo("Started recording replay");
         }
 
         public void SetStartTime()
         {
-            _startTime = new DateTimeOffset(DateTime.Now.ToUniversalTime());
-            TootTallyLogger.LogInfo($"Replay started recording at {_startTime}");
+            _replayData.SetStartTime();
+            TootTallyLogger.LogInfo($"Replay started recording at {_replayData.starttime}");
         }
 
         public void SetEndTime()
         {
-            _endTime = new DateTimeOffset(DateTime.Now.ToUniversalTime());
-            TootTallyLogger.LogInfo($"Replay recording finished at {_endTime}");
+            _replayData.SetEndTime();
+            TootTallyLogger.LogInfo($"Replay recording finished at {_replayData.endtime}");
         }
 
-        public void RecordFrameData(GameController __instance)
+        public void SetUsernameAndSongName(string username, string songname)
+        {
+            _replayData.username = username;
+            _replayData.song = songname;
+        }
+
+
+        public void RecordFrameData(GameController __instance, float time, float noteHolderPosition)
         {
             if (Input.touchCount > 0) _wasTouchScreenUsed = true;
 
-            float noteHolderPosition = __instance.noteholderr.anchoredPosition.x * GetNoteHolderPrecisionMultiplier(); // the slower the scrollspeed , the better the precision
-            float pointerPos = __instance.pointer.transform.localPosition.y * 100; // 2 decimal precision
-            float mousePosX = Input.mousePosition.x;
-            float mousePosY = Input.mousePosition.y;
+            _currentFrame = new dynamic[] { Round(time, 10f), Round(noteHolderPosition, 10f), Round(__instance.pointer.transform.localPosition.y), (int)Input.mousePosition.x, (int)Input.mousePosition.y };
 
-            _lastFrameData[(int)FrameDataStructure.NoteHolder] = (int)noteHolderPosition;
-            _lastFrameData[(int)FrameDataStructure.PointerPosition] = (int)pointerPos;
-            _lastFrameData[(int)FrameDataStructure.MousePositionX] = (int)mousePosX;
-            _lastFrameData[(int)FrameDataStructure.MousePositionY] = (int)mousePosY;
+            if (_lastFrame == null || (_currentFrame[2] != _lastFrame[2] && _currentFrame[1] != _lastFrame[1] && _currentFrame[4] != _lastFrame[4]))
+            {
+                _replayData.framedata.Add(_currentFrame);
+                _lastFrame = _currentFrame;
+            }
 
-            if (_frameData.Count < 1 || (_frameData.Last()[(int)FrameDataStructure.PointerPosition] != _lastFrameData[(int)FrameDataStructure.PointerPosition]
-                && _frameData.Last()[(int)FrameDataStructure.NoteHolder] != _lastFrameData[(int)FrameDataStructure.NoteHolder]
-                && _frameData.Last()[(int)FrameDataStructure.MousePositionY] != _lastFrameData[(int)FrameDataStructure.MousePositionY]))
-                _frameData.Add(new int[] { (int)noteHolderPosition, (int)pointerPos, (int)mousePosX, (int)mousePosY });
-           
         }
+
+        private const float FLOAT_PRECISION = 1000f;
+        public float Round(float f, float mult = 1f) => Mathf.Round(f * FLOAT_PRECISION * mult) / (FLOAT_PRECISION * mult);
 
         public void RecordNoteDataPrefix(GameController __instance)
         {
-            var noteIndex = __instance.currentnoteindex;
-            var totalScore = __instance.totalscore;
-            var multiplier = __instance.multiplier;
-            var currentHealth = __instance.currenthealth;
-            var noteScoreAverage = __instance.notescoreaverage * 1000;
-
-            _noteData.Add(new int[] { noteIndex, totalScore, multiplier, (int)currentHealth, -1, (int)noteScoreAverage });           
-        }
-
-
-        public void RecordToot(GameController __instance)
-        {
-            float noteHolderPosition = __instance.noteholderr.anchoredPosition.x * GetNoteHolderPrecisionMultiplier(); // the slower the scrollspeed , the better the precision
-
-            _tootData.Add(new int[] { (int)noteHolderPosition });
+            _currentNote = new dynamic[9];
+            _currentNote[0] = Round(__instance.notescoreaverage);
+            _currentNote[1] = __instance.released_button_between_notes ? 1 : 0;
         }
 
         public void RecordNoteDataPostfix(GameController __instance)
         {
-            _maxCombo = __instance.highestcombo_level;
-            var noteLetter = _scores_A != __instance.scores_A ? 4 :
-               _scores_B != __instance.scores_B ? 3 :
-               _scores_C != __instance.scores_C ? 2 :
-               _scores_D != __instance.scores_D ? 1 : 0;
-            _noteData[_noteData.Count - 1][(int)NoteDataStructure.NoteJudgement] = noteLetter;
-
-            _scores_A = __instance.scores_A;
-            _scores_B = __instance.scores_B;
-            _scores_C = __instance.scores_C;
-            _scores_D = __instance.scores_D;
-            _scores_F = __instance.scores_F;
-
-
+            _currentNote[2] = __instance.currentnoteindex;
+            _currentNote[3] = __instance.highestcombocounter;
+            _currentNote[4] = __instance.multiplier;
+            _currentNote[5] = __instance.totalscore;
+            _currentNote[6] = Round(__instance.currenthealth);
+            _currentNote[7] = _maxCombo = __instance.highestcombo_level;
+            _currentNote[8] = _lastTally;
+            _replayData.notedata.Add(_currentNote);
         }
 
-        public string GetRecordedReplayJson(string uuid, float targetFramerate)
+        private int _lastTally;
+
+        public void SaveLastNoteTally(int tallyIndex) => _lastTally = tallyIndex;
+
+        public void RecordToot(float time, float noteHolderPosition, bool isTooting)
         {
-            var trackRef = GlobalVariables.chosen_track_data.trackref;
-            var track = TrackLookup.lookup(trackRef);
-            string songHash;
+            _replayData.tootdata.Add(new dynamic[] { Round(time, 10f), Round(noteHolderPosition, 10f), isTooting ? 1 : 0 });
+        }
 
-            if (track is CustomTrack)
+        public string GetRecordedReplayJson(string uuid)
+        {
+            _replayData.uuid = uuid;
+            _replayData.input = GetInputTypeString();
+            _replayData.finalscore = GlobalVariables.gameplay_scoretotal;
+            _replayData.maxcombo = _maxCombo;
+            //That's absolutely disgusting I love it.
+            _replayData.finalnotetallies = new int[]
             {
-                songHash = SongDataHelper.GetSongHash(track);
-            }
-            else
-            {
-                var tmb = SongDataHelper.GenerateBaseTmb(track);
-                //FileHelper.WriteJsonToFile(Paths.BepInExRootPath, "basetmb.tmb", tmb);
-                songHash = SongDataHelper.CalcSHA256Hash(Encoding.UTF8.GetBytes(tmb));
-            }
-
-            string username = Plugin.userInfo.username;
-
-            string startDateTimeUnix = _startTime.ToUnixTimeSeconds().ToString();
-            string endDateTimeUnix = _endTime.ToUnixTimeSeconds().ToString();
-
-            string inputType = GetInputTypeString();
-            var replayJson = new SerializableClass.ReplayData();
-            replayJson.version = REPLAY_VERSION;
-            replayJson.username = username;
-            replayJson.starttime = startDateTimeUnix;
-            replayJson.endtime = endDateTimeUnix;
-            replayJson.uuid = uuid;
-            replayJson.input = inputType;
-            replayJson.song = track.trackname_long;
-            replayJson.samplerate = targetFramerate;
-            replayJson.scrollspeed = GlobalVariables.gamescrollspeed;
-            replayJson.gamespeedmultiplier = float.Parse(ReplaySystemManager.gameSpeedMultiplier.ToString("0.00"));
-            replayJson.audiolatency = GlobalVariables.localsettings.latencyadjust;
-            replayJson.gamemodifiers = GameModifierManager.GetModifiersString();
-            replayJson.pluginbuilddate = Plugin.BUILDDATE;
-            replayJson.gameversion = Application.version;
-            replayJson.songhash = songHash;
-            replayJson.finalscore = GlobalVariables.gameplay_scoretotal;
-            replayJson.maxcombo = _maxCombo;
-
-            var noteJudgmentData = new List<int>
-            {
-                _scores_A,
-                _scores_B,
-                _scores_C,
-                _scores_D,
-                _scores_F
+                _replayData.notedata.Count(x => x[(int)NDStruct.TL] == 4),
+                _replayData.notedata.Count(x => x[(int)NDStruct.TL] == 3),
+                _replayData.notedata.Count(x => x[(int)NDStruct.TL] == 2),
+                _replayData.notedata.Count(x => x[(int)NDStruct.TL] == 1),
+                _replayData.notedata.Count(x => x[(int)NDStruct.TL] == 0),
             };
 
-            replayJson.finalnotetallies = noteJudgmentData.ToArray();
-            OptimizeNoteData(ref _noteData);
-            OptimizeTootData(ref _tootData);
-            _noteData[_noteData.Count - 1][1] = GlobalVariables.gameplay_scoretotal; // Manually set the last note's totalscore to the actual totalscore because game is weird...
-
-            replayJson.framedata = _frameData;
-            replayJson.notedata = _noteData;
-            replayJson.tootdata = _tootData;
-
-            return JsonConvert.SerializeObject(replayJson);
+            return JsonConvert.SerializeObject(_replayData);
         }
         private string GetInputTypeString()
         {
@@ -198,56 +154,28 @@ namespace TootTally.Replays
                 return "Touch";
             return "Mouse";
         }
-        private static bool CheckIfSameValue(int index1, int index2, int dataIndex, List<int[]> dataList) => dataList[index1][dataIndex] == dataList[index2][dataIndex];
-
-        private static void OptimizeTootData(ref List<int[]> rawReplayTootData)
-        {
-            for (int i = 0; i < rawReplayTootData.Count - 1; i++)
-                //if two toot happens on the same frame, probably is inputFix so unsync the frames
-                if (CheckIfSameValue(i, i + 1, (int)TootDataStructure.NoteHolder, rawReplayTootData))
-                    rawReplayTootData[i][(int)TootDataStructure.NoteHolder]++;
-
-        }
-
-        private static void OptimizeNoteData(ref List<int[]> rawReplayNoteData)
-        {
-            //Current glitch in the game that duplicate a note score but doesn't have any judgement.
-            rawReplayNoteData.RemoveAll(x => x[(int)NoteDataStructure.NoteJudgement] == -1);
-        }
-
-        private static List<int[]> GetDuplicatesFromDataList(float valueToFind, int dataIndex, List<int[]> dataList) => dataList.FindAll(data => data[dataIndex] == valueToFind);
         #endregion
 
         #region ReplayPlayer
 
         public void OnReplayPlayerStart()
         {
-            _lastTiming = 0;
-            _totalScore = 0;
             _frameIndex = 0;
             _tootIndex = 0;
-            _replayNoteTally = new int[5];
             _isTooting = false;
+            _isLastNote = false;
+            _currentNote = _replayData.notedata.First();
+            _currentToot = new dynamic[] { 0, 0, 0 };
+            _currentFrame = _replayData.framedata.First();
         }
 
-        public void OnReplayRewind(float noteHolderNewPosX, GameController __instance)
+        public void OnReplayRewind(float newTiming, GameController __instance)
         {
-            float currentMapPosition = (noteHolderNewPosX / 45f) * GetNoteHolderPrecisionMultiplier(); //45f is a rough estimate of the ratio between localPos and worldPos
-
-            _frameIndex = Mathf.Clamp(_frameData.FindIndex(frame => frame[(int)FrameDataStructure.NoteHolder] <= currentMapPosition) - 1, 0, _frameData.Count - 1);
-            _lastTiming = _frameIndex > 0 ? _frameData[_frameIndex - 1][(int)FrameDataStructure.NoteHolder] : 0;
-            _tootIndex = Mathf.Clamp(_tootData.FindIndex(frame => frame[(int)TootDataStructure.NoteHolder] <= currentMapPosition) - 1, 0, _tootData.Count - 1);
-            _isTooting = _tootIndex % 2 == 1; //if the count is odd then its tooting.
+            _frameIndex = Mathf.Clamp(_replayData.framedata.FindIndex(frame => (float)frame[(int)FDStruct.T] > newTiming) - 1, 0, _replayData.framedata.Count - 1);
+            _tootIndex = Mathf.Clamp(_replayData.tootdata.FindIndex(frame => (float)frame[(int)TDStruct.T] > newTiming) - 1, 0, _replayData.tootdata.Count - 1);
 
             if (__instance.currentnoteindex != 0)
-                __instance.currentscore = _noteData.Find(note => note[(int)NoteDataStructure.NoteIndex] == __instance.currentnoteindex - 1)[(int)NoteDataStructure.TotalScore];
-        }
-
-
-        public void OnReplayPlayerStop()
-        {
-            GlobalVariables.gameplay_notescores = _finalNoteTally;
-            GlobalVariables.gameplay_scoretotal = _finalTotalScore;
+                __instance.currentscore = (int)_replayData.notedata.Find(note => (int)note[(int)NDStruct.I] == __instance.currentnoteindex - 1)[(int)NDStruct.S];
         }
 
         public ReplayState LoadReplay(string replayFileName)
@@ -266,102 +194,92 @@ namespace TootTally.Replays
 
             string jsonFileFromZip = FileHelper.ReadJsonFromFile(replayDir, replayFileName + ".ttr");
 
-            var replayJson = JsonConvert.DeserializeObject<SerializableClass.ReplayData>(jsonFileFromZip);
-            if (incompatibleReplayPluginBuildDate.Contains(replayJson.pluginbuilddate.ToString()))
+            _replayData = JsonConvert.DeserializeObject<ReplayData>(jsonFileFromZip);
+            if (incompatibleReplayVersions.Contains(_replayData.pluginbuilddate.ToString()))
             {
-                PopUpNotifManager.DisplayNotif($"Replay incompatible:\nReplay Build Date is {replayJson.pluginbuilddate}\nCurrent Build Date is {Plugin.BUILDDATE}", GameTheme.themeColors.notification.errorText);
+                PopUpNotifManager.DisplayNotif($"Replay incompatible:\nReplay Build Date is {_replayData.pluginbuilddate}\nCurrent Build Date is {Plugin.BUILDDATE}", GameTheme.themeColors.notification.errorText);
                 TootTallyLogger.LogError("Cannot load replay:");
-                TootTallyLogger.LogError("   Replay Build Date is " + replayJson.pluginbuilddate);
+                TootTallyLogger.LogError("   Replay Build Date is " + _replayData.pluginbuilddate);
                 TootTallyLogger.LogError("   Current Plugin Build Date " + Plugin.BUILDDATE);
                 return ReplayState.ReplayLoadErrorIncompatible;
             }
-            GlobalVariables.gamescrollspeed = replayJson.scrollspeed;
-            _frameData = replayJson.framedata;
-            _noteData = replayJson.notedata;
-            _tootData = replayJson.tootdata;
-            _finalNoteTally = replayJson.finalnotetallies;
-            _finalTotalScore = replayJson.finalscore;
-            _replayUsername = replayJson.username;
-            _replaySong = replayJson.song;
-            _replaySpeed = replayJson.gamespeedmultiplier != 0 ? replayJson.gamespeedmultiplier : 1f;
-            GameModifierManager.LoadModifiersFromString(replayJson.gamemodifiers ?? "");
-            _replaybuilddate = replayJson.pluginbuilddate;
+            GlobalVariables.gamescrollspeed = _replayData.scrollspeed;
+            GameModifierManager.LoadModifiersFromString(_replayData.gamemodifiers ?? "");
 
             return ReplayState.ReplayLoadSuccess;
         }
 
-        public void PlaybackReplay(GameController __instance)
+        public void PlaybackReplay(GameController __instance, float time)
         {
             Cursor.visible = true;
             if (!__instance.controllermode) __instance.controllermode = true; //Still required to not make the mouse position update
 
-            //Backward compatibility with old replays
-            var currentMapPosition = (_replaybuilddate < 20230705 ?
-                 __instance.noteholder.transform.position.x : __instance.noteholderr.anchoredPosition.x) * GetNoteHolderPrecisionMultiplier();
+            PlaybackFrameData(time);
+            PlaybackTootData(time);
 
-            __instance.totalscore = _totalScore;
-            if (_frameData.Count > _frameIndex && _lastPosition != 0)
-                InterpolateCursorPosition(currentMapPosition, __instance);
-
-            PlaybackFrameData(currentMapPosition, __instance);
-            PlaybackTootData(currentMapPosition, __instance);
+            if (_replayData.framedata.Count > _frameIndex && _lastFrame != null && _currentFrame != null)
+                InterpolateCursorPosition(time, __instance);
         }
 
-        private void InterpolateCursorPosition(float currentMapPosition, GameController __instance)
+        private void InterpolateCursorPosition(float time, GameController __instance)
         {
-            var newCursorPosition = EasingHelper.Lerp(_lastPosition, _nextPositionTarget, (_lastTiming - currentMapPosition) / (_lastTiming - _nextTimingTarget));
-            SetCursorPosition(__instance, newCursorPosition);
-            __instance.puppet_humanc.doPuppetControl(-newCursorPosition / 225); //225 is half of the Gameplay area:450
-        }
-
-        private void PlaybackFrameData(float currentMapPosition, GameController __instance)
-        {
-            if (_frameData.Count > _frameIndex && currentMapPosition <= _frameData[_frameIndex][(int)FrameDataStructure.NoteHolder])
+            if (_currentFrame[(int)FDStruct.T] - _lastFrame[(int)FDStruct.T] > 0)
             {
-                _lastTiming = _frameData[_frameIndex][(int)FrameDataStructure.NoteHolder];
-                _lastPosition = _frameData[_frameIndex][(int)FrameDataStructure.PointerPosition] / 100f;
-            }
-
-            while (_frameData.Count > _frameIndex && currentMapPosition <= _frameData[_frameIndex][(int)FrameDataStructure.NoteHolder]) //smaller or equal to because noteholder goes toward negative
-            {
-
-                if (_frameIndex < _frameData.Count - 1)
-                {
-                    _nextTimingTarget = _frameData[_frameIndex + 1][(int)FrameDataStructure.NoteHolder];
-                    _nextPositionTarget = _frameData[_frameIndex + 1][(int)FrameDataStructure.PointerPosition] / 100f;
-                }
-                else
-                {
-                    _nextTimingTarget = _lastTiming;
-                    _nextPositionTarget = _lastPosition;
-                }
-
-                SetCursorPosition(__instance, _frameData[_frameIndex][(int)FrameDataStructure.PointerPosition] / 100f);
-
-                _frameIndex++;
+                var newCursorPosition = EasingHelper.Lerp((float)_lastFrame[(int)FDStruct.P], (float)_currentFrame[(int)FDStruct.P], (float)((time - (float)_lastFrame[(int)FDStruct.T]) / ((float)_currentFrame[(int)FDStruct.T] - (float)_lastFrame[(int)FDStruct.T])));
+                SetCursorPosition(__instance, newCursorPosition);
+                __instance.puppet_humanc.doPuppetControl(-newCursorPosition / 225); //225 is half of the Gameplay area:450
             }
         }
 
-        private void PlaybackTootData(float currentMapPosition, GameController __instance)
+        private void PlaybackFrameData(float time)
         {
-            if (_tootData.Count > _tootIndex && currentMapPosition <= _tootData[_tootIndex][(int)TootDataStructure.NoteHolder])
+
+            if (_lastFrame != _currentFrame && time >= _currentFrame[(int)FDStruct.T])
+                _lastFrame = _currentFrame;
+
+            if (_replayData.framedata.Count > _frameIndex && (_currentFrame == null || time >= _currentFrame[(int)FDStruct.T]))
             {
-                _isTooting = !_isTooting;
-                _tootIndex++;
+                _frameIndex = _replayData.framedata.FindIndex(_frameIndex > 1 ? _frameIndex - 1 : 0, x => time < x[(int)FDStruct.T]);
+                if (_replayData.framedata.Count > _frameIndex && _frameIndex != -1)
+                    _currentFrame = _replayData.framedata[_frameIndex];
             }
         }
 
-        public void SetNoteScore(GameController __instance)
+        private void PlaybackTootData(float time)
         {
-            var note = _noteData.Find(x => x[(int)NoteDataStructure.NoteIndex] == __instance.currentnoteindex);
+            if (time >= _currentToot[(int)TDStruct.T] && _isTooting != (_currentToot[(int)TDStruct.O] == 1))
+                _isTooting = _currentToot[(int)TDStruct.O] == 1;
 
-            if (note != null)
+            if (_replayData.tootdata.Count > _tootIndex && time >= _currentToot[(int)TDStruct.T])
+                _currentToot = _replayData.tootdata[_tootIndex++];
+        }
+
+        public void SetNoteScorePrefix(GameController __instance)
+        {
+            if (!_isLastNote)
             {
-                __instance.totalscore = _totalScore = note[(int)NoteDataStructure.TotalScore]; //total score has to be set postfix as well because notes SOMEHOW still give more points than they should during replay...
-                __instance.multiplier = note[(int)NoteDataStructure.Multiplier];
-                __instance.currenthealth = note[(int)NoteDataStructure.CurrentHealth];
-                int tallyIndex = Mathf.Clamp(note[(int)NoteDataStructure.NoteJudgement], 0, 4); //Temporary fix for note tally being -1 sometimes?
-                _replayNoteTally[tallyIndex]++;
+                _currentNote = _replayData.notedata.Find(x => x[(int)NDStruct.I] == __instance.currentnoteindex);
+                _isLastNote = _replayData.notedata.Last()[(int)NDStruct.I] == __instance.currentnoteindex;
+            }
+            if (_currentNote != null)
+            {
+                __instance.notescoreaverage = (float)_currentNote[(int)NDStruct.NS];
+                __instance.released_button_between_notes = _currentNote[(int)NDStruct.R] == 1;
+            }
+        }
+
+        public void SetNoteScorePostFix(GameController __instance)
+        {
+            if (_currentNote != null)
+            {
+                __instance.rainbowcontroller.champmode = _currentNote[(int)NDStruct.H] == 100;
+                __instance.multiplier = (int)_currentNote[(int)NDStruct.M];
+                if (__instance.currentscore < 0)
+                    __instance.currentscore = (int)_currentNote[(int)NDStruct.S];
+                __instance.totalscore = (int)_currentNote[(int)NDStruct.S];
+                __instance.currenthealth = (float)_currentNote[(int)NDStruct.H];
+                __instance.highestcombo_level = (int)_currentNote[(int)NDStruct.HC];
+                _currentNote = null;
             }
         }
 
@@ -377,40 +295,93 @@ namespace TootTally.Replays
         public void ClearData()
         {
             TootTallyLogger.LogInfo("Replay data cleared");
-            _frameData.Clear();
-            _noteData.Clear();
-            _tootData.Clear();
+            _replayData?.ClearData();
         }
 
-        public void SetUsernameAndSongName(string username, string songname)
+        [Serializable]
+        private class ReplayData
         {
-            _replayUsername = username;
-            _replaySong = songname;
+            public string version { get; set; }
+            public string username { get; set; }
+            public string starttime { get; set; }
+            public string endtime { get; set; }
+            public string input { get; set; }
+            public string song { get; set; }
+            public string uuid { get; set; }
+            public int samplerate { get; set; }
+            public float scrollspeed { get; set; }
+            public float gamespeedmultiplier { get; set; }
+            public string gamemodifiers { get; set; }
+            public float audiolatency { get; set; }
+            public int pluginbuilddate { get; set; }
+            public string gameversion { get; set; }
+            public string songhash { get; set; }
+            public int finalscore { get; set; }
+            public int maxcombo { get; set; }
+            public int[] finalnotetallies { get; set; }
+            public List<dynamic[]> framedata { get; set; }
+            public List<dynamic[]> notedata { get; set; }
+            public List<dynamic[]> tootdata { get; set; }
+
+            public ReplayData(int sampleRate)
+            {
+                framedata = new List<dynamic[]>();
+                notedata = new List<dynamic[]>();
+                tootdata = new List<dynamic[]>();
+                pluginbuilddate = Plugin.BUILDDATE;
+                version = REPLAY_VERSION;
+                username = Plugin.userInfo.username;
+                audiolatency = GlobalVariables.localsettings.latencyadjust;
+                gameversion = Application.version;
+                samplerate = sampleRate;
+            }
+
+            public void ClearData()
+            {
+                framedata?.Clear();
+                notedata?.Clear();
+                tootdata?.Clear();
+            }
+
+            public void SetStartTime()
+            {
+                starttime = new DateTimeOffset(DateTime.Now.ToUniversalTime()).ToUnixTimeSeconds().ToString();
+            }
+
+            public void SetEndTime()
+            {
+                endtime = new DateTimeOffset(DateTime.Now.ToUniversalTime()).ToUnixTimeSeconds().ToString();
+            }
+
         }
 
-        public static float GetNoteHolderPrecisionMultiplier() => 10 / (GlobalVariables.gamescrollspeed <= 1 ? GlobalVariables.gamescrollspeed : 1);
-
-        private enum FrameDataStructure
+        public enum FDStruct
         {
-            NoteHolder = 0,
-            PointerPosition = 1,
-            MousePositionX = 2,
-            MousePositionY = 3,
+            T,
+            N,
+            P,
+            MX,
+            MY
         }
 
-        private enum TootDataStructure
+        public enum TDStruct
         {
-            NoteHolder = 0,
+            T,
+            N,
+            O
         }
 
-        private enum NoteDataStructure
+        public enum NDStruct
         {
-            NoteIndex = 0,
-            TotalScore = 1,
-            Multiplier = 2,
-            CurrentHealth = 3,
-            NoteJudgement = 4,
-            NoteScore = 5,
+            NS, //NoteScoreAverage
+            R, //ReleasedBetweenNotes
+            I, //Index
+            C, //Combo
+            M, //Multiplier
+            S, //Score
+            H, //Health
+            HC, //HighestCombo
+            TL, //TallyScore
         }
 
         public enum ReplayState
